requirements.txt
streamlit
openai
requests
pandas
bcrypt
Pillow
